Add files via upload
Customers began to leave the Bank every month. A little, but noticeable. Bank marketers decided that it was cheaper to retain existing customers than to attract new ones. We need to predict whether the client will leave the bank in the near future or not. We are provided with historical data on customer behavior and termination of agreements with the bank.